# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ProviEmplea EVA3",
        "description": "API para gestión de candidatos con CV ciego, empresas, relación empresa-candidato y salud de la API.\n\n**Rate Limiting:** 60 requests/minuto por IP.\n**Caché:** Endpoints GET cacheados por 5 minutos.\n**Tiempo de respuesta objetivo:** < 200ms.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local Docker"
        }
    ],
    "paths": {
...
        "/empresas": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Listar empresas",
                "description": "Obtiene todas las empresas registradas en la plataforma.",
                "operationId": "22631014ece23858d19a7ef1c89f0e78",
                "responses": {
                    "200": {
                        "description": "Lista de empresas obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Empresa"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Crear empresa",
                "description": "Registra una nueva empresa en la plataforma.",
                "operationId": "edd395f6abdaa55cb0284b3b198d7ad6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Empresa"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empresa creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empresa"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/empresas/{id}": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Obtener empresa por ID",
                "description": "Obtiene una empresa específica mediante su ID.",
                "operationId": "1ab6ccd465b58367f4064aa417a103be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la empresa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empresa"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Actualizar empresa",
                "description": "Actualiza los datos de una empresa registrada.",
                "operationId": "b61661cfb086a49afd6ebebf15ed6011",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la empresa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Empresa"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empresa"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Eliminar empresa",
                "description": "Elimina una empresa por ID.",
                "operationId": "56b2aac859ceb472597b45436c7f5f12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la empresa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa eliminada correctamente"
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/empresas/{id}/estado": {
            "patch": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Activar o desactivar empresa",
                "description": "Permite encender o apagar el estado de una empresa.",
                "operationId": "ce6108b8c25f010a6a62a4fd5ed70b50",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la empresa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "estado"
                                ],
                                "properties": {
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empresa"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Empresa": {
                "title": "Empresa",
                "required": [
                    "rut_empresa",
                    "nombre_empresa",
                    "rubro",
                    "email_contacto"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "rut_empresa": {
                        "type": "string",
                        "example": "76.123.456-7"
                    },
                    "nombre_empresa": {
                        "type": "string",
                        "example": "Tech Solutions"
                    },
                    "rubro": {
                        "type": "string",
                        "example": "Tecnología"
                    },
                    "email_contacto": {
                        "type": "string",
                        "example": "rrhh@techsolutions.cl"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "+56222222222"
                    },
                    "beneficios": {
                        "type": "string",
                        "example": "Seguro complementario, trabajo remoto"
                    },
                    "estado": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}